test: extract pmem2 utils for libpmem2 source from libpmem2 config utils
