+++ src/test/pmem2_integration/Makefile
@@ -11,7 +11,8 @@ vpath %.c $(TOP)/src/test/unittest
 TARGET = pmem2_integration
 OBJS = pmem2_integration.o\
 	ut_pmem2_utils.o\
-	ut_pmem2_config.o
+	ut_pmem2_config.o\
+	ut_pmem2_source.o
 
 LIBPMEM2=y
