+++ src/test/pmem2_integration/pmem2_integration.vcxproj
@@ -69,12 +69,13 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\unittest\ut_pmem2_config.c" />
+    <ClCompile Include="..\unittest\ut_pmem2_source.c" />
     <ClCompile Include="..\unittest\ut_pmem2_utils.c" />
     <ClCompile Include="pmem2_integration.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\unittest\unittest.h" />
-    <ClInclude Include="..\unittest\ut_pmem2_utils.h" />
+    <ClInclude Include="..\unittest\ut_pmem2.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\libpmem2\libpmem2.vcxproj">
