+++ src/test/pmem2_map/Makefile
@@ -13,7 +13,8 @@ INCS += -I$(TOP)/src/libpmem2
 TARGET = pmem2_map
 OBJS += pmem2_map.o\
 	ut_pmem2_config.o\
-	ut_pmem2_utils.o
+	ut_pmem2_utils.o\
+	ut_pmem2_source.o
 
 LIBPMEM2=internal-debug
