+++ src/test/pmem2_mem_ext/pmem2_mem_ext.vcxproj.filters
@@ -23,5 +23,8 @@
     <ClCompile Include="pmem2_mem_ext.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\unittest\ut_pmem2_source.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
