+++ src/test/pmem2_memcpy/pmem2_memcpy.vcxproj.filters
@@ -29,9 +29,15 @@
     <ClCompile Include="..\unittest\ut_pmem2_utils.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\unittest\ut_pmem2_source.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="memcpy_common.h">
+      <Filter>Header files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\unittest\ut_pmem2.h">
       <Filter>Header files</Filter>
     </ClInclude>
   </ItemGroup>
