+++ src/test/pmem2_memset/pmem2_memset.c
@@ -9,8 +9,7 @@
 
 #include "unittest.h"
 #include "file.h"
-#include "ut_pmem2_utils.h"
-#include "ut_pmem2_config.h"
+#include "ut_pmem2.h"
 #include "memset_common.h"
 
 static void
