+++ src/test/pmem2_movnt_align/pmem2_movnt_align.vcxproj
@@ -61,6 +61,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\unittest\ut_pmem2_config.c" />
+    <ClCompile Include="..\unittest\ut_pmem2_source.c" />
     <ClCompile Include="..\unittest\ut_pmem2_utils.c" />
     <ClCompile Include="movnt_align_common.c" />
     <ClCompile Include="pmem2_movnt_align.c" />
@@ -77,6 +78,7 @@
     <None Include="TESTS.py" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\unittest\ut_pmem2.h" />
     <ClInclude Include="movnt_align_common.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
