+++ src/test/pmem2_source_alignment/pmem2_source_alignment.vcxproj.filters
@@ -17,9 +17,6 @@
     <ClCompile Include="..\..\libpmem2\errormsg.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\libpmem2\config.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="..\unittest\ut_pmem2_config.c">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -35,21 +32,21 @@
     <ClCompile Include="..\..\libpmem2\source_windows.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\unittest\ut_pmem2_source.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\unittest\ut_pmem2_config.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="..\unittest\ut_pmem2_utils.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\libpmem2\config.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\libpmem2\pmem2_utils.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\libpmem2\pmem2.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\libpmem2\source.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\unittest\ut_pmem2.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
