+++ src/test/pmem2_source_size/pmem2_source_size.c
@@ -9,8 +9,7 @@
 
 #include "fault_injection.h"
 #include "unittest.h"
-#include "ut_pmem2_utils.h"
-#include "ut_pmem2_config.h"
+#include "ut_pmem2.h"
 #include "ut_fh.h"
 #include "config.h"
 #include "out.h"
