+++ src/test/unittest/ut_pmem2.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: BSD-3-Clause
-/* Copyright 2019, Intel Corporation */
+/* Copyright 2019-2020, Intel Corporation */
 
 /*
  * ut_pmem2.h -- utility helper functions for libpmem tests
@@ -9,6 +9,7 @@
 #define UT_PMEM2_H 1
 
 #include "ut_pmem2_config.h"
+#include "ut_pmem2_source.h"
 #include "ut_pmem2_utils.h"
 
 #endif /* UT_PMEM2_H */
